+++ stroom-dashboard-client/src/main/java/stroom/dashboard/client/table/cf/ExpressionMatcher.java
@@ -34,8 +34,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 public class ExpressionMatcher {
     private static final String DELIMITER = ",";
@@ -44,8 +42,10 @@ public class ExpressionMatcher {
     private final Map<String, String> fieldNameToJsDateFormat;
 
     public ExpressionMatcher(final List<Field> fields) {
-        this.fieldNameToFieldMap = fields.stream()
-                .collect(Collectors.toMap(Field::getName, Function.identity()));
+        this.fieldNameToFieldMap = new HashMap<>();
+        for (final Field field : fields) {
+            fieldNameToFieldMap.putIfAbsent(field.getName(), field);
+        }
 
         // For any date cols with a format str, convert the string to js moment syntax
         this.fieldNameToJsDateFormat = new HashMap<>();
@@ -62,7 +62,7 @@ public class ExpressionMatcher {
                         .getSettings())
                         .getPattern())
                         .ifPresent(format ->
-                                fieldNameToJsDateFormat.put(field.getName(), format)));
+                                fieldNameToJsDateFormat.putIfAbsent(field.getName(), format)));
     }
 
     public boolean match(final Map<String, Object> attributeMap, final ExpressionItem item) {
@@ -377,8 +377,8 @@ public class ExpressionMatcher {
             if (value instanceof String) {
                 String valueStr = (String) value;
                 try {
-                        // This is a term value so will be IDO format
-                        return ClientDateUtil.fromISOString(valueStr);
+                    // This is a term value so will be IDO format
+                    return ClientDateUtil.fromISOString(valueStr);
                 } catch (final NumberFormatException e) {
                     GWT.log("Unable to parse a date/time from value \"" + valueStr + "\"");
                     throw new MatchException(
@@ -408,7 +408,7 @@ public class ExpressionMatcher {
 
     private Long[] getDates(final Object value) {
         if (value == null) {
-           return new Long[0];
+            return new Long[0];
         } else {
             final String[] values = value.toString().split(DELIMITER);
             final Long[] dates = new Long[values.length];